# backend/src/main/resources/openapi/sims-openapi.yaml
openapi: 3.0.3
info:
  title: SIMS API — Exam, Fee and Communication modules
  description: |
    School Information Management System — the exam, fee and communication modules
    after the business-workflow redesign.

    ## Conventions

    **Nested business objects.** Responses embed the data a screen needs rather than
    bare foreign keys: `subject`, `student`, `schoolClass`, `createdBy` and
    `updatedBy` are objects carrying ids *and* names. The old flat `*Id` fields are
    still returned but are deprecated and will be removed once the React app has
    migrated off them.

    **Pagination.** List endpoints return a Spring `Page`:
    `?page=0&size=20&sort=createdAt,desc`.

    **Actor fields.** Requests carry `actorId` (admin staff) or `enteredById` /
    `updatedById` (teacher) because authentication is still being built by another
    developer. These fields disappear once the JWT principal is available.

    **Marks and fee "sheets".** Marks and fee statuses are worked on a section at a
    time, so each has a `GET /sheet` that returns the whole roster pre-filled and a
    `PUT /sheet` that saves it in one transaction.
  version: 2.0.0
  contact:
    name: SIMS Team
servers:
  - url: http://localhost:8081
    description: Local development
  - url: https://api.example.com
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/exams/{id}:
    parameters:
      - $ref: '#/components/parameters/PathId'
    get:
      tags: [Exams]
      summary: Get exam by id
      operationId: getExamById
      responses:
        '200':
          description: Exam
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExamResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    put:
      tags: [Exams]
      summary: Update an exam
      description: |
        Rejected once the exam is `PUBLISHED`. The status field is ignored here —
        status only changes through the publish endpoint.
      operationId: updateExam
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExamRequest' }
      responses:
        '200':
          description: Exam updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExamResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Exam is published, or the name duplicates another exam in the year
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    delete:
      tags: [Exams]
      summary: Delete an exam
      description: |
        Rejected when the exam is published, when marks have been recorded against
        it, or while subjects are still mapped to it.
      operationId: deleteExam
      responses:
        '204': { description: Exam deleted }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Published, has marks, or still has subjects mapped
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
  /api/v1/exam-subjects:
    get:
      tags: [Exam Subjects]
      summary: List exam subjects
      operationId: getAllExamSubjects
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Page of exam subjects
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageExamSubjectResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    post:
      tags: [Exam Subjects]
      summary: Map a subject into an exam
      operationId: createExamSubject
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExamSubjectRequest' }
      responses:
        '200':
          description: Exam subject created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExamSubjectResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Already mapped for this class, or the exam is published
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/exam-subjects/{id}:
    parameters:
      - $ref: '#/components/parameters/PathId'
    get:
      tags: [Exam Subjects]
      summary: Get exam subject by id
      operationId: getExamSubjectById
      responses:
        '200':
          description: Exam subject
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExamSubjectResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    put:
      tags: [Exam Subjects]
      summary: Update an exam subject
      description: |
        Re-mapping is rejected once marks exist, and `maxMarks` cannot be lowered
        below the highest mark already recorded.
      operationId: updateExamSubject
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExamSubjectRequest' }
      responses:
        '200':
          description: Exam subject updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExamSubjectResponse' }
        '400':
          description: Validation failure, or maxMarks below a recorded mark
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Exam published, marks already recorded, or duplicate mapping
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    delete:
      tags: [Exam Subjects]
      summary: Delete an exam subject
      description: Rejected when the exam is published or marks have been recorded.
      operationId: deleteExamSubject
      responses:
        '204': { description: Exam subject deleted }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Exam published, or marks already recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/exam-subjects/exam/{examId}:
    parameters:
      - name: examId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Exam Subjects]
      summary: Subjects mapped to an exam
      description: Bounded by the exam, so returned unpaginated.
      operationId: getExamSubjectsByExamId
      responses:
        '200':
          description: Exam subjects
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ExamSubjectResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }

  # ==================================================================
  # Exam Timetable
  # ==================================================================
  /api/v1/exam-timetables:
    get:
      tags: [Exam Timetable]
      summary: List timetable entries
      operationId: getAllExamTimetables
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Page of timetable entries
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageExamTimetableResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    post:
      tags: [Exam Timetable]
      summary: Schedule an exam subject
      description: |
        One slot per exam subject (enforced by a unique index). The date must fall
        inside the exam's academic year, and the exam must not be published.
      operationId: createExamTimetable
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExamTimetableRequest' }
      responses:
        '200':
          description: Timetable entry created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExamTimetableResponse' }
        '400':
          description: Validation failure, or date outside the academic year
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Subject already scheduled, or the exam is published
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/exam-timetables/{id}:
    parameters:
      - $ref: '#/components/parameters/PathId'
    get:
      tags: [Exam Timetable]
      summary: Get timetable entry by id
      operationId: getExamTimetableById
      responses:
        '200':
          description: Timetable entry
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExamTimetableResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    put:
      tags: [Exam Timetable]
      summary: Reschedule an exam subject
      operationId: updateExamTimetable
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExamTimetableRequest' }
      responses:
        '200':
          description: Timetable entry updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExamTimetableResponse' }
        '400':
          description: Validation failure, or date outside the academic year
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Subject already scheduled, or the exam is published
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    delete:
      tags: [Exam Timetable]
      summary: Delete a timetable entry
      operationId: deleteExamTimetable
      responses:
        '204': { description: Timetable entry deleted }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Exam is published
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/exam-timetables/exam/{examId}:
    parameters:
      - name: examId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Exam Timetable]
      summary: Full schedule of an exam
      description: Ordered by date then time.
      operationId: getExamTimetablesByExamId
      responses:
        '200':
          description: Timetable entries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ExamTimetableResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }

  # ==================================================================
  # Marks — workflow
  # ==================================================================
...
  /api/v1/marks/exam-subject/{examSubjectId}:
    parameters:
      - name: examSubjectId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Marks]
      summary: Marks for an exam subject
      description: Prefer `GET /api/v1/marks/sheet` for the entry screen.
      operationId: getMarksByExamSubject
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Page of marks
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageMarkResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
components:
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 20 }
    Sort:
      name: sort
      in: query
      required: false
      description: '`property,(asc|desc)`; repeatable.'
      schema:
        type: array
        items: { type: string }
      example: ['createdAt,desc']
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by the login endpoint. Authentication is still being built, so
        endpoints are currently open and role rules are recorded as TODOs beside each
        controller method.
  responses:
...
    NotFound:
      description: A referenced record does not exist
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            timestamp: '2026-08-02T10:30:00Z'
            status: 404
            error: Not Found
            message: 'Exam not found with id: 9f1c8b2e-...'
            path: /api/v1/exams/9f1c8b2e-...
    ValidationError:
      description: Bean validation failure; offending fields are listed in `fieldErrors`
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            timestamp: '2026-08-02T10:30:00Z'
            status: 400
            error: Bad Request
            message: Validation failed
            path: /api/v1/exams
            fieldErrors:
              examName: Exam name is required
              actorId: Acting admin id is required
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ExamStatus:
      type: string
      enum: [DRAFT, PUBLISHED]
...
    SubjectType:
      type: string
      enum: [COMPULSORY, ELECTIVE]

    # ---------------- shared summaries ----------------
    ActorSummary:
      type: object
      description: A person who created or last changed a record.
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: R. Iyer }
        role:
          type: string
          description: Designation, e.g. PRINCIPAL or SENIOR_TEACHER.
          example: SENIOR_TEACHER
    ClassSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Grade 10 }
    SubjectSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Mathematics }
        code: { type: string, example: MATH001 }
        type: { $ref: '#/components/schemas/SubjectType' }
    AcademicYearSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        label: { type: string, example: 2025-2026 }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
...
    StudentSummary:
      type: object
      description: A student seen through their enrollment.
      properties:
        enrollmentId: { type: string, format: uuid }
        studentId: { type: string, format: uuid }
        name: { type: string, example: Aarav Sharma }
        rollNumber: { type: string, example: '04' }
        admissionNumber: { type: string, example: ADM-2025-0142 }

    # ---------------- errors & paging ----------------
    ErrorResponse:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer, example: 409 }
        error: { type: string, example: Conflict }
        message: { type: string }
        path: { type: string }
        fieldErrors:
          type: object
          nullable: true
          additionalProperties: { type: string }
          description: Present only for bean validation failures.
      required: [timestamp, status, error, message, path]
    PageMetadata:
      type: object
      properties:
        totalElements: { type: integer, format: int64, example: 137 }
        totalPages: { type: integer, example: 7 }
        number: { type: integer, description: Zero-based page index, example: 0 }
        size: { type: integer, example: 20 }
        numberOfElements: { type: integer, example: 20 }
        first: { type: boolean }
        last: { type: boolean }
        empty: { type: boolean }

    # ---------------- exam ----------------
    ExamSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Term 1 Examination }
        status: { $ref: '#/components/schemas/ExamStatus' }
    ExamRequest:
      type: object
      required: [academicYearId, examName, actorId]
      properties:
        academicYearId: { type: string, format: uuid }
        examName: { type: string, maxLength: 255, example: Term 1 Examination }
        status:
          allOf: [{ $ref: '#/components/schemas/ExamStatus' }]
          description: Optional; defaults to DRAFT. Ignored on update.
        actorId:
          type: string
          format: uuid
          description: |
            Admin staff performing the action — becomes `createdBy` on create and
            `updatedBy` on update. Removed once authentication lands.
    ExamResponse:
      type: object
      properties:
        examId: { type: string, format: uuid }
        examName: { type: string }
        status: { $ref: '#/components/schemas/ExamStatus' }
        academicYear: { $ref: '#/components/schemas/AcademicYearSummary' }
        createdBy: { $ref: '#/components/schemas/ActorSummary' }
        updatedBy: { $ref: '#/components/schemas/ActorSummary' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        academicYearId:
          type: string
          format: uuid
          deprecated: true
          description: Use `academicYear`.
        createdById:
          type: string
          format: uuid
          deprecated: true
          description: Use `createdBy`.
    ExamSubjectRequest:
      type: object
      required: [examId, subjectId, classId, maxMarks, actorId]
      properties:
        examId: { type: string, format: uuid }
        subjectId: { type: string, format: uuid }
        classId: { type: string, format: uuid }
        maxMarks: { type: integer, minimum: 1, maximum: 1000, example: 100 }
        actorId: { type: string, format: uuid }
    ExamSubjectResponse:
      type: object
      properties:
        examSubjectId: { type: string, format: uuid }
        maxMarks: { type: integer, example: 100 }
        exam: { $ref: '#/components/schemas/ExamSummary' }
        subject: { $ref: '#/components/schemas/SubjectSummary' }
        schoolClass: { $ref: '#/components/schemas/ClassSummary' }
        updatedBy: { $ref: '#/components/schemas/ActorSummary' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        examId: { type: string, format: uuid, deprecated: true }
        subjectId: { type: string, format: uuid, deprecated: true }
        classId: { type: string, format: uuid, deprecated: true }
    ExamTimetableRequest:
      type: object
      required: [examSubjectId, examDate, examTime, actorId]
      properties:
        examSubjectId: { type: string, format: uuid }
        examDate: { type: string, format: date, example: '2026-09-15' }
        examTime: { type: string, example: '09:30:00' }
        actorId: { type: string, format: uuid }
    ExamTimetableResponse:
      type: object
      properties:
        examTimetableId: { type: string, format: uuid }
        examDate: { type: string, format: date }
        examTime: { type: string, example: '09:30:00' }
        maxMarks: { type: integer }
        exam: { $ref: '#/components/schemas/ExamSummary' }
        subject: { $ref: '#/components/schemas/SubjectSummary' }
        schoolClass: { $ref: '#/components/schemas/ClassSummary' }
        updatedBy: { $ref: '#/components/schemas/ActorSummary' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        examSubjectId: { type: string, format: uuid, deprecated: true }
...
    MarkResponse:
      type: object
      properties:
        markId: { type: string, format: uuid }
        marksObtained: { type: number, format: double, example: 85.5 }
        maxMarks: { type: integer, example: 100 }
        student: { $ref: '#/components/schemas/StudentSummary' }
        subject: { $ref: '#/components/schemas/SubjectSummary' }
        exam: { $ref: '#/components/schemas/ExamSummary' }
        enteredBy: { $ref: '#/components/schemas/ActorSummary' }
        updatedBy: { $ref: '#/components/schemas/ActorSummary' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        enrollmentId: { type: string, format: uuid, deprecated: true }
        examSubjectId: { type: string, format: uuid, deprecated: true }
        enteredById: { type: string, format: uuid, deprecated: true }
...
    PageExamSubjectResponse:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/ExamSubjectResponse' }
    PageExamTimetableResponse:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/ExamTimetableResponse' }
    PageMarkResponse:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/MarkResponse' }
...